Use unsignedBigInteger for tender_id in migrations
Change tender_id columns from unsignedInteger to unsignedBigInteger across multiple migration files to align with big primary keys and avoid potential integer overflow. Indexes and unique constraints are preserved.

Updated files:
- database/migrations/2027_02_24_000001_create_jawatankuasas_table.php
- database/migrations/2027_04_06_000002_create_jawatankuasa_perolehan_meetings_table.php
- database/migrations/2027_04_07_000001_create_perakuan_jabatan_kertas_taklimat_tables.php
- database/migrations/2027_04_07_000002_create_perakuan_jabatan_pengesyoran_pembekal_table.php
- database/migrations/2027_04_07_000003_create_jawatankuasa_perolehan_kertas_keputusans_table.php
- database/migrations/2027_04_13_000001_create_jawatankuasa_perolehan_pemilihan_tables.php

+++ database/migrations/2027_02_24_000001_create_jawatankuasas_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         }
         Schema::create('jawatankuasas', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('tender_id')->nullable();
+            $table->unsignedBigInteger('tender_id')->nullable();
             $table->enum('jenis_jawatankuasa', ['spec', 'open', 'tech', 'fin'])->index();
             $table->enum('p_p', ['1', '0'])->index();
             $table->enum('peranan', ['1', '2', '3'])->index();
